Test for area_cvx as a float (#97)
tests: test for area_cvx as a float, consistent with output from Shape-In (#96)

+++ tests/test_rtdc_write_hdf5.py
@@ -306,7 +306,7 @@ def test_real_time():
     masks = np.zeros((M, shy, shx), dtype=np.bool_)
     traces = {"fl1_median": np.arange(M * 55).reshape(M, 55)}
     axis1 = np.linspace(0, 1, M)
-    axis2 = np.arange(M)
+    axis2 = np.arange(float(M))
     rtdc_file = tempfile.mktemp(suffix=".rtdc",
                                 prefix="dclab_test_realtime_")
     with h5py.File(rtdc_file, "w") as fobj:
@@ -335,7 +335,7 @@ def test_real_time():
         assert events["contour"]["0"].shape == (10, 2)
         assert events["trace"]["fl1_median"].shape == (N, 55)
         assert np.dtype(events["area_um"]) == float
-        assert np.dtype(events["area_cvx"]) == int
+        assert np.dtype(events["area_cvx"]) == float
 
 
 def test_real_time_single():
@@ -354,7 +354,7 @@ def test_real_time_single():
         # simulate real time and write one image at a time
         for ii in range(N):
             data = {"area_um": ii * .1,
-                    "area_cvx": ii * 5,
+                    "area_cvx": ii * 5.,
                     "image": image * ii,
                     "contour": contour,
                     "mask": mask,
@@ -373,7 +373,7 @@ def test_real_time_single():
         assert events["contour"]["0"].shape == (11, 2)
         assert events["trace"]["fl1_median"].shape == (N, 43)
         assert np.dtype(events["area_um"]) == float
-        assert np.dtype(events["area_cvx"]) == int
+        assert np.dtype(events["area_cvx"]) == float
         logs = rtdc_data["logs"]
         assert len(logs["log1"]) == N
